Промежуточный коммит по корзине.

+++ commerce/interaction-api/src/main/java/ru/yandex/practicum/model/cart/ChangeProductQuantityRequest.java
@@ -12,5 +12,5 @@ import java.util.UUID;
 @Builder
 public class ChangeProductQuantityRequest {
     private UUID productId;
-    private Integer newQuantity;
+    private Long  newQuantity;
 }
